# backend/src/swagger-output.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Newsx API",
    "description": "API Documentation for Newsx application, a platform for creating and managing blog posts, comments, likes, and user interactions.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://blog-sphere-backend-ruby.vercel.app",
      "description": "Production server"
    }
  ],
  "basePath": "/",
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter token in format: Bearer <token>"
    }
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "statusCode": { "type": "integer", "example": 200 },
        "data": { "type": "object", "nullable": true },
        "message": { "type": "string", "example": "Operation successful" },
        "success": { "type": "boolean", "example": true }
      }
    },
...
  },
  "paths": {
...
    "/api/v1/posts/{id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get a single post",
        "description": "Retrieves a post by its ID.",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the post",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Post retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/ApiResponse" },
                "example": {
                  "statusCode": 200,
                  "data": {
                    "_id": "60d21b4667d0d8992e610c85",
                    "title": "My First Post",
                    "content": "This is the content of my post."
                  },
                  "message": "Post retrieved successfully",
                  "success": true
                }
              }
            }
          },
          "404": { "description": "Post not found" }
        }
      },
      "patch": {
        "tags": ["Posts"],
        "summary": "Update a post",
        "description": "Updates an existing post with optional image upload. Requires authentication and ownership.",
        "security": [{ "BearerAuth": [] }],
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the post to update",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Title of the post"
                  },
                  "content": {
                    "type": "string",
                    "description": "Content of the post"
                  },
                  "category": {
                    "type": "string",
                    "description": "Category of the post"
                  },
                  "tags": {
                    "type": "string",
                    "description": "Comma-separated tags"
                  },
                  "contentTable": {
                    "type": "string",
                    "description": "Content table for the post"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional post image (e.g., PNG, JPEG)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/ApiResponse" },
                "example": {
                  "statusCode": 200,
                  "data": {
                    "_id": "60d21b4667d0d8992e610c85",
                    "title": "Updated Post",
                    "content": "This is the updated content."
                  },
                  "message": "Post updated successfully",
                  "success": true
                }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Unauthorized (not post owner)" },
          "404": { "description": "Post not found" },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "Delete a post",
        "description": "Deletes a post by its ID, including associated comments and likes. Requires authentication and ownership.",
        "security": [{ "BearerAuth": [] }],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the post to delete",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/ApiResponse" },
                "example": {
                  "statusCode": 200,
                  "data": null,
                  "message": "Post deleted successfully",
                  "success": true
                }
              }
            }
          },
          "403": { "description": "Unauthorized (not post owner)" },
          "404": { "description": "Post not found" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  },
...
}